build(bazel): align root rust targets to library+test pattern
