+++ BUILD.bazel
@@ -1,12 +1,17 @@
 load("@crate_index//:defs.bzl", "aliases", "all_crate_deps")
-load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_test")
+load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_library", "rust_test")
 
 package(default_visibility = ["//visibility:public"])
 
-rust_binary(
-    name = "agenthub",
-    srcs = glob(["src/**/*.rs"]),
-    crate_root = "src/main.rs",
+rust_library(
+    name = "agenthub_lib",
+    srcs = glob(
+        ["src/**/*.rs"],
+        exclude = ["src/main.rs"],
+    ),
+    crate_name = "agenthub",
+    crate_root = "src/lib.rs",
+    compile_data = glob(["web/dist/**"]),
     edition = "2024",
     aliases = aliases(),
     deps = all_crate_deps(normal = True) + [
@@ -18,27 +23,24 @@ rust_binary(
 
 rust_test(
     name = "agenthub_unit_tests",
-    srcs = glob(["src/**/*.rs"]),
-    crate_root = "src/main.rs",
+    crate = ":agenthub_lib",
     edition = "2024",
     aliases = aliases(
-        normal = True,
-        normal_dev = True,
-        proc_macro = True,
         proc_macro_dev = True,
-    ),
-    deps = all_crate_deps(
-        normal = True,
         normal_dev = True,
-    ) + [
-        "//crates/agenthub-acp:agenthub_acp",
-        "//crates/agenthub-team-actor:agenthub_team_actor",
-    ],
-    proc_macro_deps = all_crate_deps(
-        proc_macro = True,
-        proc_macro_dev = True,
     ),
+    deps = all_crate_deps(normal_dev = True),
+    proc_macro_deps = all_crate_deps(proc_macro_dev = True),
     timeout = "long",
+)
+
+rust_binary(
+    name = "agenthub",
+    srcs = ["src/main.rs"],
+    edition = "2024",
+    aliases = aliases(),
+    deps = all_crate_deps(normal = True) + [":agenthub_lib"],
+    proc_macro_deps = all_crate_deps(proc_macro = True),
 )
 
 rust_test(
